requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
